controller/requirements.txt
kubernetes
PyYAML
pytz
